test(advisory): widen SymfonyProcessComposerAuditRunner timeout to float for a faster test
Mirrors the same speed-up already applied to ProcessGitChangedFilesResolver:
timeoutSeconds now accepts a fractional value, letting the timeout test use
0.1s instead of waiting a full second for a real process timeout.
AdvisorySourceUnavailableException::forTimeout()'s message format switches
from %d to %s to render the fractional value without distorting the
integer-second case real users see.

+++ src/Audit/Infrastructure/Advisory/SymfonyProcessComposerAuditRunner.php
@@ -38,7 +38,7 @@ final readonly class SymfonyProcessComposerAuditRunner implements ComposerAuditR
      * @param ?Closure(string): Process $processBuilder defaults to the standard composer-audit command; tests inject a stub
      */
     public function __construct(
-        private int $timeoutSeconds = self::DEFAULT_TIMEOUT_SECONDS,
+        private float $timeoutSeconds = self::DEFAULT_TIMEOUT_SECONDS,
         private ?Closure $processBuilder = null,
     ) {}
 
@@ -60,7 +60,7 @@ final readonly class SymfonyProcessComposerAuditRunner implements ComposerAuditR
         $process = $builder($projectPath);
 
         try {
-            $process->setTimeout((float) $this->timeoutSeconds);
+            $process->setTimeout($this->timeoutSeconds);
             $process->run();
         } catch (ProcessTimedOutException $processTimedOutException) {
             throw AdvisorySourceUnavailableException::forTimeout($this->timeoutSeconds, $processTimedOutException);
